+++ app/Http/Controllers/KeuanganController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Keuangan;
+use App\Models\Masjid;
+use Illuminate\Support\Facades\Gate;
 use Auth;
 
 
@@ -14,10 +16,15 @@ class KeuanganController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['show']);
+    }
+
     public function index()
     {
-        $listKeuangan = Keuangan::get();
-        return view('keuangan.daftar_keuangan', ['listKeuangan' => $listKeuangan]);
+        
     }
 
     /**
@@ -27,6 +34,10 @@ class KeuanganController extends Controller
      */
     public function create()
     {
+        $masjid = Masjid::find(Auth::user()->id);
+        if (! Gate::allows('isMyAccount', $masjid)) {
+            abort(403);
+        }
         return view('keuangan.form_buat_keuangan');
     }
 
@@ -59,7 +70,7 @@ class KeuanganController extends Controller
 
         $keuangan->save();
 
-        return redirect('/keuangan');
+        return redirect('keuangan/'. Auth::user()->id);
 
     }
 
@@ -71,7 +82,10 @@ class KeuanganController extends Controller
      */
     public function show($id)
     {
-        //
+        $masjid = Masjid::find($id);
+        $listKeuangan = Keuangan::where('id_masjid','=', $id)->get();
+        
+        return view('keuangan.daftar_keuangan', ['listKeuangan' => $listKeuangan, 'masjid' => $masjid]);
     }
 
     /**
@@ -82,6 +96,10 @@ class KeuanganController extends Controller
      */
     public function edit($id)
     {
+        $keuangan = Keuangan::find($id);
+        if (! Gate::allows('isMyAccount', $keuangan->masjid)) {
+            abort(403);
+        }
         $keuangan = Keuangan::find($id);
         return view('keuangan.form_edit_keuangan', ['keuangan' => $keuangan]);
     }
@@ -115,7 +133,7 @@ class KeuanganController extends Controller
 
         $keuangan->save();
 
-        return redirect('/keuangan');
+        return redirect('keuangan/'. Auth::user()->id);
 
     }
 
@@ -128,8 +146,13 @@ class KeuanganController extends Controller
     public function destroy($id)
     {
         $keuangan = Keuangan::find($id);
+        if (! Gate::allows('isMyAccount', $keuangan->masjid)) {
+            abort(403);
+        }
+            
         $keuangan->delete();
 
-        return redirect('/keuangan');
+        return redirect('keuangan/'.Auth::user()->id);
+        
     }
 }
